Move dialog directory and fix paths
