+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/area-of-influence-form-model.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
 var template = require('./area-of-influence-form.tpl');
 var DistanceConverter = require('builder/value-objects/distance-converter');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 
 var AREA_OF_INFLUENCE_TYPES = require('./area-of-influence-types');
 var TYPE_TO_META_MAP = {};
