+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/base-analysis-form-model.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var Backbone = require('backbone');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var camshaftReference = require('builder/data/camshaft-reference');
 
 /**
