+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/closest-form-model.js
@@ -1,6 +1,6 @@
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
 var ColumnOptions = require('builder/editor/layers/layer-content-views/analyses/column-options');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var template = require('./closest-form.tpl');
 
 var MAX_RESPONSES = 32;
