+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
 var template = require('./connect-with-lines.tpl');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var ColumnOptions = require('builder/editor/layers/layer-content-views/analyses/column-options');
 
 var CONNECT_WITH_LINES = require('./connect-with-lines-types');
