+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/data-observatory-multiple-measure-model.js
@@ -5,7 +5,7 @@ var templateList = require('./data-observatory-multiple-measure-list.tpl');
 var templateItem = require('./data-observatory-multiple-measure-list-item.tpl');
 var NestedMeasurement = require('./data-observatory-multiple-nested-measure-model');
 var DataObservatoryRegions = require('builder/data/data-observatory/regions-collection');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var camshaftReference = require('builder/data/camshaft-reference');
 var itemRegionListTemplate = require('./data-observatory-multiple-region-list-item.tpl');
