+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/data-observatory-multiple-nested-measure-model.js
@@ -4,7 +4,7 @@ var _ = require('underscore');
 var DataObservatoryNormalize = require('builder/data/data-observatory/normalize-collection');
 var DataObservatoryTimespan = require('builder/data/data-observatory/timespan-collection');
 var DataObservatoryBoundaries = require('builder/data/data-observatory/boundaries-collection');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var MeasurementModel = require('builder/data/data-observatory/measurement-model');
 var DataObservatoryColumnName = require('builder/data/data-observatory/column-name');
 var checkAndBuildOpts = require('builder/helpers/required-opts');
