+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/deprecated-sql-function-form-model.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var cdb = require('@carto/carto.js');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var checkAndBuildOpts = require('builder/helpers/required-opts');
 var availableFunctionsQuery = require('./available-functions.tpl');
 var template = require('./deprecated-sql-function.tpl');
