+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/fallback-form-model.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var camshaftReference = require('builder/data/camshaft-reference');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var ARRAY_SPLIT_COMBO = ',';
 
 /**
