+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/filter-intersection-form-model.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
 var template = require('./filter-intersection-form.tpl');
 var ColumnOptions = require('builder/editor/layers/layer-content-views/analyses/column-options');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 /**
  *  Filter points in polygon model
  */
