+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/georeference-form-model.js
@@ -3,7 +3,7 @@ var BaseAnalysisFormModel = require('./base-analysis-form-model');
 var template = require('./georeference-form.tpl');
 var ColumnOptions = require('builder/editor/layers/layer-content-views/analyses/column-options');
 var GEOREFERENCE_TYPES = require('./georeference-types');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 
 var TYPE_TO_META_MAP = {};
 GEOREFERENCE_TYPES.map(function (d) {
