+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/group-points-form-model.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var template = require('./group-points.tpl');
 var ColumnOptions = require('builder/editor/layers/layer-content-views/analyses/column-options');
