+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/merge-form-model.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
 var template = require('./merge-form.tpl');
 var ColumnOptions = require('builder/editor/layers/layer-content-views/analyses/column-options');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var MERGE_FIELDS = 'join_operator,left_source_column,right_source_column,source_geometry_selector,left_source_columns,right_source_columns';
 
 /**
