+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/analyses/analysis-form-models/moran-form-model.js
@@ -1,5 +1,5 @@
 var BaseAnalysisFormModel = require('./base-analysis-form-model');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var template = require('./moran-form.tpl');
 var ColumnOptions = require('builder/editor/layers/layer-content-views/analyses/column-options');
