+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/legend/form/legend-bubble-definition-form-model.js
@@ -1,5 +1,5 @@
 var _ = require('underscore');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var LegendBaseDefModel = require('./legend-base-definition-form-model');
 var LegendColorHelper = require('./legend-color-helper');
