+++ lib/assets/javascripts/builder/editor/layers/layer-content-views/legend/form/legend-nested-category-definition-form-model.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var Backbone = require('backbone');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 var LegendColors = require('./legend-color-range');
 
 var defaultFill = {
