+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/labels-placement.js
@@ -1,5 +1,5 @@
 var _ = require('underscore');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 
 module.exports = {
   generate: function () {
