+++ lib/assets/test/spec/builder/editor/layers/layer-content-view/analyses/analyses-form-models/closest-form-model.spec.js
@@ -3,7 +3,7 @@ var cdb = require('@carto/carto.js');
 var Model = require('builder/editor/layers/layer-content-views/analyses/analysis-form-models/closest-form-model');
 var BaseAnalysisFormModel = require('builder/editor/layers/layer-content-views/analyses/analysis-form-models/base-analysis-form-model');
 var analyses = require('builder/data/analyses');
-var DialogConstants = require('builder/components/form-components/editors/_constants/_dialogs');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 
 describe('editor/layers/layer-content-views/analyses/analysis-form-models/closest-form-model', function () {
   var cdbSQLBackup;
